BM.circle {animation}	R Documentation

Brownian Motion in a circle

Description

Several points moving randomly in a circle.

Usage

BM.circle(n = 20, col = rainbow(n), ...)

Arguments

n
number of points

col
colors of points

...
other parameters passed to points

Details

This is a solution to the question raised in R-help: https://stat.ethz.ch/pipermail/r-help/2008-December/183018.html.

Value

Invisible NULL.

Note

The maximum number of steps in the motion is specified in ani.options('nmax').

Author(s)

Yihui Xie

References

Examples at https://yihui.org/animation/example/bm-circle/

See Also

brownian.motion, rnorm

[Package animation version 2.8 Index]