ani.start(animation)	R Documentation

Start the generation of an HTML animation page

Description

Copy JavaScript file ‘FUN.js’ and CSS file ‘ANI.css’ to the same directory as the HTML animation page, create a directory ‘images’ and open a graphics device in this directory (the device is specified as ani.dev in ani.options). The prompt of the current R session is modified (by default ANI> ).

Usage


ani.start(...)

Arguments

... 	 arguments passed to ani.options to set animation parameters

Value

None (invisible `NULL').

Note

After calling ani.start, either animation functions in this package or R script of your own can be used to generate & save animated pictures using proper graphics devices (specified as ani.dev in ani.options), then watch your animation by ani.stop().

Note that former image files in the directory ‘images’ will be removed.

Author(s)

Yihui Xie

References

http://animation.yihui.name/animation:create_html_animation_page

See Also

ani.options, ani.stop

Examples


## Not run: 
 
# save the animation in HTML pages and auto-browse it 
ani.options(ani.width = 600, ani.height = 500, interval = 0.2)
ani.start()
boot.iid()
ani.stop()

## End(Not run)

[Package animation version 0.2-1 Index]