ani.stop(animation)	R Documentation

Write the HTML animation page

Description

Write the HTML animation page, restore previous options such as prompt and close the graphical device opened in ani.start.

Usage


ani.stop()

Value

None (invisible `NULL'); a string will be printed in the console indicating where is the HTML file.

Note

The content of the HTML file completely depends on the parameters set in ani.options.

Author(s)

Yihui Xie

References

http://animation.yihui.name/animation:create_html_animation_page

See Also

ani.options, ani.start

Examples


## Not run: 
 
# save the animation in HTML pages and auto-browse it
ani.options(ani.width = 600, ani.height = 500, interval = 0.2)
ani.start()
boot.iid()
ani.stop()

## End(Not run)

[Package animation version 0.2-1 Index]