boot.lowess {animation}	R Documentation

Bootstrapping with LOWESS

Description

Sample the original data with replacement and fit LOWESS curves accordingly.

Usage

boot.lowess(x, y = NULL, f = 2/3, iter = 3, line.col = "#FF000033", ...)

Arguments

x, y, f, iter
passed to lowess

line.col
the color of the LOWESS lines

...
other arguments passed to the scatterplot by plot

Details

We keep on resampling the data and finally we will see several bootstrapped LOWESS curves, which may give us a rough idea about a “confidence interval” of the LOWESS fit.

Author(s)

Yihui Xie

References

Examples at https://yihui.org/animation/example/boot-lowess/

[Package animation version 2.8 Index]