brownian.motion {animation}	R Documentation

Demonstration of Brownian motion on the 2D plane

Description

Brownian motion, or random walk, can be regarded as the trace of some cumulative normal random numbers.

Usage

brownian.motion(n = 10, xlim = c(-20, 20), ylim = c(-20, 20), ...)

Arguments

n
Number of points in the scatterplot

xlim, ylim
Arguments passed to plot.default to control the apperance of the scatterplot (title, points, etc), see points for details.

...
other arguments passed to plot.default

Details

The location of the next step is “current location + random Gaussian numbers”, i.e.,

x[k + 1] = x[k] + rnorm(1)

y[k + 1] = y[k] + rnorm(1)

where (x, y) stands for the location of a point.

Value

None (invisible NULL).

Note

The maximum number of steps in the motion is specified in ani.options('nmax').

Author(s)

Yihui Xie

References

Examples at https://yihui.org/animation/example/brownian-motion/

See Also

rnorm

[Package animation version 2.8 Index]