highlight.def(animation)	R Documentation

Create R definition file for the software Highlight

Description

The default definition file for R in Highlight is somewhat incomplete, and this function is to dynamically generate such a file according to packages in the search path.

Usage


highlight.def(file = "r.lang")

Arguments

file 	 the path of the output definition file.

Details

First all the functions are listed out by ls; then some constants and operators are removed from this long list; at last these characters are written into the ‘file’.

Value

None.

Author(s)

Yihui Xie

References

Highlight by Andre Simon: http://www.andre-simon.de/

See Also

ls, cat

Examples


# generate the definition file in getwd() 
highlight.def()

# include functions in package 'animation'
library(animation) 
highlight.def()
 

[Package animation version 1.0-0 Index]