price.ani {animation}	R Documentation

Demonstrate stock prices in animations

Description

This function can display the frequencies of stock prices in a certain time span with the span changing.

Usage

price.ani(
  price,
  time,
  time.begin = min(time),
  span = 15 * 60,
  ...,
  xlab = "price",
  ylab = "frequency",
  xlim,
  ylim,
  main
)

Arguments

price
stock prices

time
time corresponding to prices

time.begin
the time for the animation to begin (default to be the minimum time)

span
time span (unit in seconds; default to be 15 minutes)

...
other arguments passed to plot

xlab, ylab, xlim, ylim, main
they are passed to plot with reasonable default values

Value

invisible NULL

Author(s)

Yihui Xie

References

Examples at https://yihui.org/animation/example/price-ani/

[Package animation version 2.8 Index]