qpdf {animation}	R Documentation

A wrapper for the PDF toolkit qpdf

Description

If the tool qpdf is available in the system, it will be called to manipulate the PDF files (especially to compress the PDF files).

Usage

qpdf(input, output, options = "--stream-data=compress")

Arguments

input
the path of the input PDF file

output
the path of the output (if missing, output will be the same as input)

options
options for qpdf (default to be '--stream-data=compress', i.e. compress the PDF files)

Details

This is a wrapper to call qpdf. The path of qpdf should be set via ani.options(qpdf = 'path/to/qpdf').

See the reference for detailed usage of qpdf.

Value

if ani.options('qpdf') is non-NULL, then this function returns the status of the operation (0 for success; see system); otherwise a warning will be issued

Author(s)

Yihui Xie

References

Examples at https://yihui.org/animation/example/qpdf/

qpdf official website https://qpdf.sourceforge.io/

[Package animation version 2.8 Index]