sample.cluster(animation)	R Documentation

Demonstration for cluster sampling

Description

Every rectangle stands for a cluster, and the simple random sampling without replacement is performed for each cluster. All points in the clusters being sampled will be drawn out.

Usage


sample.cluster(pop = ceiling(10 * runif(10, 0.2, 1)), size = 3,
    p.col = c("blue", "red"), p.cex = c(1, 3), ...)

Arguments

pop 	 a vector for the size of each cluster in the population.

size 	 the number of clusters to be drawn out.

p.col, p.cex 	 different colors / magnification rate to annotate the population and the sample

... 	 other arguments passed to rect to annotate the ``clusters''

Value

None (invisible `NULL').

Author(s)

Yihui Xie

References

Cochran, W G (1977) Sampling Techniques, Wiley, ISBN 0-471-16240-X

http://animation.yihui.name/samp:cluster_sampling

See Also

sample

Examples



oopt = ani.options(interval = 1, nmax = 30)
op = par(mar = rep(1, 4))
sample.cluster(col = c("bisque", "white"))
par(op)
## Not run:  

# HTML animation page
ani.options(ani.height = 350, ani.width = 500, outdir = getwd(), nmax = 30,
    interval = 1, title = "Demonstration of the cluster sampling",
    description = "Once a cluster is sampled, all its elements will be
    chosen.")
ani.start()
par(mar = rep(1, 4), lwd = 2)
sample.cluster(col = c("bisque", "white"))
ani.stop()

## End(Not run)
ani.options(oopt)

[Package animation version 1.0-1 Index]