sample.simple {animation}	R Documentation

Demonstration for simple random sampling without replacement

Description

The whole sample frame is denoted by a matrix (nrow * ncol) in the plane just for convenience, and the points being sampled are marked out (by red circles by default). Each member of the population has an equal and known chance of being selected.

Usage

sample.simple(
  nrow = 10,
  ncol = 10,
  size = 15,
  p.col = c("blue", "red"),
  p.cex = c(1, 3)
)

Arguments

nrow
the desired number of rows of the sample frame.

ncol
the desired number of columns of the sample frame.

size
the sample size.

p.col, p.cex
different colors /magnification rate to annotate the population and the sample

Value

None (invisible NULL).

Author(s)

Yihui Xie

References

Examples at https://yihui.org/animation/example/sample-simple/

See Also

sample, sample.ratio, sample.cluster, sample.strat, sample.system

[Package animation version 2.8 Index]