sample.strat {animation}	R Documentation

Demonstration for the stratified sampling

Description

Each rectangle stands for a stratum, and the simple random sampling without replacement is performed within each stratum. The points being sampled are marked out (by red circles by default).

Usage

sample.strat(
  pop = ceiling(10 * runif(10, 0.5, 1)),
  size = ceiling(pop * runif(length(pop), 0, 0.5)),
  p.col = c("blue", "red"),
  p.cex = c(1, 3),
  ...
)

Arguments

pop
a vector for the size of each stratum in the population.

size
a corresponding vector for the sample size in each stratum (recycled if necessary).

p.col, p.cex
different colors /magnification rate to annotate the population and the sample

...
other arguments passed to rect to annotate the “strata”

Value

None (invisible 'NULL').

Author(s)

Yihui Xie

References

Examples at https://yihui.org/animation/example/sample-strat/

See Also

sample, sample.simple, sample.cluster, sample.ratio, sample.system

[Package animation version 2.8 Index]