sample.system(animation)	R Documentation

Demonstration for systematic sampling

Description

The whole sample frame is denoted by a matrix (nrow * ncol) in the plane, and the sample points with equal intervals are drawn out according to a random starting point. The points being sampled are marked by red circles.

Usage


sample.system(nrow = 10, ncol = 10, size = 15,
    p.col = c("blue", "red"), p.cex = c(1, 3))

Arguments

nrow 	 the desired number of rows of the sample frame.

ncol 	 the desired number of columns of the sample frame.

size 	 the sample size.

p.col, p.cex 	 different colors / magnification rate to annotate the population and the sample

Value

None (invisible `NULL').

Author(s)

Yihui Xie

References

Cochran, W G (1977) Sampling Techniques, Wiley, ISBN 0-471-16240-X

See Also

sample

Examples


oopt = ani.options(interval = 1, nmax = 30)
op = par(mar = rep(1, 4), lwd = 2)
sample.system()
par(op)
## Not run: 

# HTML animation page
ani.options(ani.height = 350, ani.width = 500, outdir = getwd(), nmax = 30,
    interval = 1, title = "Demonstration of the systematic sampling",
    description = "Sampling with equal distances.")
ani.start()
par(mar = rep(1, 4), lwd = 2)
sample.system()
ani.stop()

## End(Not run)
ani.options(oopt)

[Package animation version 1.0-1 Index]