vanke1127 {animation}	R Documentation

Stock prices of Vanke Co., Ltd on 2009/11/27

Description

This is a sample of stock prices of the Vanke Co., Ltd on 2009/11/27.

Format

A data frame with 2831 observations on the following 2 variables.

time

POSIXt: the time corresponding to stock prices

price

a numeric vector: stock prices

Source

This data can be obtained from most stock websites.

[Package animation version 2.8 Index]