vi.lilac.chaser {animation}	R Documentation

Visual Illusions: Lilac Chaser

Description

Stare at the center cross for a few (say 30) seconds to experience the phenomena of the illusion.

Usage

vi.lilac.chaser(np = 16, col = "magenta", bg = "gray", p.cex = 7, c.cex = 5)

Arguments

np
number of points

col
color of points

bg
background color of the plot

p.cex
magnification of points

c.cex
magnification of the center cross

Details

Just try it out.

Value

NULL

Note

In fact, points in the original version of ‘Lilac Chaser’ are blurred, which is not implemented in this function.

Author(s)

Yihui Xie

References

Examples at https://yihui.org/animation/example/vi-lilac.chaser/

See Also

points

[Package animation version 2.8 Index]