CLELAL09 {animation}	R Documentation

The NBA game between CLE Cavaliers and LAL Lakers on Dec 25, 2009

Description

Cleveland Cavaliers played against Los Angeles Lakers at Staples Center in LA on Dec 25, 2009 and won the game by 102:87. This data recorded the locations of players on the court and the results of the shots.

Format

A data frame with 455 observations on the following 7 variables.

player

a character vector: the current player

time

a character vector: the time

period

a numeric vector: the period (1 - 4)

realx

a numeric vector: the x-axis location

realy

a numeric vector: the y-axis location

result

a factor with levels made and missed

team

a factor with levels CLE, LAL and OFF

Note

We view the court with CLE in the left and LAL in the right: realx is the distance to the left border of CLE's court, and realy is the distance to the bottom border of the court; notice that the size of the court is 94 x 50 (feet).

Source

http://www.basketballgeek.com/data/ (transformed based on the original data)

Examples

## see demo('CLEvsLAL', package = 'animation') for a `replay' of the game

[Package animation version 2.4 Index]