ani.pause {animation}	R Documentation

Pause for a while and flush the current graphical device

Description

If this function is called in an interactive graphics device, it will pause for a time interval (by default specified in ani.options('interval')) and flush the current device; otherwise it will do nothing.

Usage

ani.pause(interval = ani.options("interval"))

Arguments

interval
a time interval to pause (in seconds)

Value

Invisible NULL.

Author(s)

Yihui Xie

See Also

dev.interactive, Sys.sleep, dev.flush

Examples

## pause for 2 seconds
oopt = ani.options(interval = 2)

for (i in 1:5) {
    plot(runif(10), ylim = c(0, 1))
    ani.pause()
}

ani.options(oopt)

## see demo('Xmas2', package = 'animation') for another example

[Package animation version 2.4 Index]