A Gallery of Animations in Statistics and Utilities to Create Animations

This package contains a variety functions for animations in statistics which could probably aid in teaching statistics and data analysis; it also has several utilities to export R animations to other formats.

Details

This package mainly makes use of HTML & JavaScript and R windows graphics devices (such as x11) to demonstrate animations in statistics; other kinds of output such as Flash (SWF) or GIF animations or PDF animations are also available if required software packages have been installed. See below for details on each type of animation.

Note

Bug reports and feature requests can be sent to https://github.com/yihui/animation/issues.

On-screen Animations

It's natural and easy to create an animation in R using the windows graphics device, e.g. in x11() or windows(). A basic scheme is like the Example 1 (see below).

On-screen animations do not depend on any third-party software, but the rendering speed of the windows graphics devices is often slow, so the animation might not be smooth (especially under Linux and Mac OS).

HTML Pages

The generation of HTML animation pages does not rely on any third-party software either, and we only need a web browser to watch the animation. See saveHTML.

The HTML interface is just like a movie player -- it comes with a series of buttons to control the animation (play, stop, next, previous, ...).

This HTML approach is flexible enough to be used even in Rweb, which means we do not really have to install R to create animations! There is a demo in system.file('misc', 'Rweb', 'demo.html', package = 'animation'). We can use saveHTML to create animations directly in Rweb; this can be helpful when we do not have R or cannot install R.

GIF Animations

If ImageMagick or GraphicsMagick has been installed, we can use im.convert or gm.convert to create a GIF animation (combining several R plots together), or use saveGIF to create a GIF animation from an R code chunk.

Flash Animations

If SWF Tools has been installed, we can use saveSWF to create a Flash animation (again, combining R plots).

PDF Animations

If LaTeX is present in the system, we can use saveLatex to insert animations into a PDF document and watch the animation using the Adobe reader.

The animation is created by the LaTeX package animate.

Video

The function saveVideo can use FFmpeg to convert images to various video formats (e.g. mp4, avi and wmv, etc).

References

Examples at https://yihui.org/animation/example/animation-package/

The associated website for this package: https://yihui.org/animation

Yihui Xie and Xiaoyue Cheng. animation: A package for statistical animations. R News, 8(2):23--27, October 2008. URL: https://CRAN.R-project.org/doc/Rnews/Rnews_2008-2.pdf

(NB: some functions mentioned in the above article have been slightly modified; see the help pages for the up-to-date usage.)

Yihui Xie (2013). animation: An R Package for Creating Animations and Demonstrating Statistical Methods. Journal of Statistical Software, 53(1), 1-27. URL doi:10.18637/jss.v053.i01 .

See also

Author

Yihui Xie