kfcv(animation)	R Documentation

Sample sizes for k-fold cross-validation

Description

Compute sample sizes for k-fold cross-validation.

Usage


kfcv(k, N)

Arguments

k 	 number of groups.

N 	 total sample size.

Details

If N/k is an integer, the sample sizes are k `N/k's (N/k, N/k, ...), otherwise the remainder will be allocated to each group as `uniformly' as possible, and at last these sample sizes will be permuted randomly.

Value

A vector of length k containing k sample sizes.

Author(s)

Yihui Xie

See Also

cv.ani

Examples


# divisible
kfcv(5, 25)

# not divisible 
kfcv(10,77) 

[Package animation version 1.0-1 Index]