pageview {animation}	R Documentation

Page views from Sep 21, 2007 to Dec 2, 2007 of Yihui's website

Description

The data is collected by Awstats for the website http://yihui.name.

Format

A data frame with 73 observations on the following 5 variables.

day

Date starts from Sep 21, 2007 to Dec 2, 2007.

visits

number of visits: a new visit is defined as each new incoming visitor (viewing or browsing a page) who was not connected to the site during last 60 min.

pages

number of times a page of the site is viewed (sum for all visitors for all visits). This piece of data differs from “files” in that it counts only HTML pages and excludes images and other files.

files

number of times a page, image, file of the site is viewed or downloaded by someone.

bandwidth

amount of data downloaded by all pages, images and files within the site (units in MegaBytes).

Source

http://yihui.name

Examples

plot(pageview[, 1:2], type = "b", col = "red", main = "Number of Visits in Yihui's Web")
## partial auto-correlation
pacf(pageview$visits)

[Package animation version 2.4 Index]