pageview(animation)	R Documentation

Data of page view from Sep 21 2007 for Yihui's website

Description

Page view data for Yihui's website from Sep 21, 2007 to a recent date.

Usage

data(pageview)

Format

A data frame with 73 observations on the following 5 variables.

day

Date starts from Sep 21, 2007 to a recent date.

visits

number of visits: a new visit is defined as each new incoming visitor (viewing or browsing a page) who was not connected to the site during last 60 min.

pages

number of times a page of the site is viewed (sum for all visitors for all visits). This piece of data differs from ``files'' in that it counts only HTML pages and excludes images and other files.

files

number of times a page, image, file of the site is viewed or downloaded by someone.

bandwidth

amount of data downloaded by all pages, images and files within the site (units in MegaBytes).

Details

The data is collected by Awstats for the website http://www.yihui.name.

Source

http://www.yihui.name/cgi-bin/awstats/awstats.pl?month=10&year=2007&output=main&config=yihuiname&framename=index

Examples


data(pageview)
plot(pageview[,1:2], type = "b", col = "red", 
  main = "Number of Visits in Yihui's Web")
# partial auto-correlation 
pacf(pageview$visits) 

[Package animation version 0.1-1 Index]