pdftk {animation}	R Documentation

A wrapper for the PDF toolkit Pdftk

Description

If the toolkit Pdftk is available in the system, it will be called to manipulate the PDF files (especially to compress the PDF files).

Usage

pdftk(input, operation = NULL, output, other.opts = "compress dont_ask")

Arguments

input
the path of the input PDF file(s)

operation
the operation to be done on the input (default to be NULL)

output
the path of the output (if missing and input is a scalar, output will be the same as input)

other.opts
other options (default to be 'compress dont_ask', i.e. compress the PDF files and do not ask the user for any input)

Details

This is a wrapper to call pdftk. The path of pdftk should be set via ani.options(pdftk = 'path/to/pdftk').

See the reference for detailed usage of pdftk.

Value

if ani.options('pdftk') is non-NULL, then this function returns the status of the operation (0 for success; see system); otherwise a warning will be issued

Author(s)

Yihui Xie

References

http://www.pdflabs.com/tools/pdftk-the-pdf-toolkit/

Examples

pdf("huge-plot.pdf")
plot(rnorm(50000))
dev.off()

## Windows
ani.options(pdftk = "D:/Installer/pdftk.exe")
pdftk("huge-plot.pdf", output = "huge-plot0.pdf")

## Linux (does not work??)
ani.options(pdftk = "pdftk")
pdftk("huge-plot.pdf", output = "huge-plot1.pdf")

ani.options(pdftk = NULL)

file.info(c("huge-plot.pdf", "huge-plot0.pdf", "huge-plot1.pdf"))["size"]

[Package animation version 2.4 Index]