pollen {animation}	R Documentation

Synthetic dataset about the geometric features of pollen grains

Description

There are 3848 observations on 5 variables. From the 1986 ASA Data Exposition dataset, made up by David Coleman of RCA Labs

Format

A data frame with 3848 observations on the following 5 variables.

RIDGE

a numeric vector

NUB

a numeric vector

CRACK

a numeric vector

WEIGHT

a numeric vector

DENSITY

a numeric vector

Source

collected from Statlib Datasets Archive: http://lib.stat.cmu.edu/data-expo/

Examples

## some dense points in the center?
plot(pollen[, 1:2], pch = 20, col = rgb(0, 0, 0, 0.1))

## see demo('pollen', package = 'animation') for a 3D demo; truth is there!

[Package animation version 2.4 Index]