price.ani {animation}	R Documentation

Demonstrate stock prices in animations

Description

This function can display the frequencies of stock prices in a certain time span with the span changing.

Usage

price.ani(price, time, time.begin = min(time), span = 15 * 60, ..., xlab = "price", 
    ylab = "frequency", xlim, ylim, main)

Arguments

price
stock prices

time
time corresponding to prices

time.begin
the time for the animation to begin (default to be the minimum time)

span
time span (unit in seconds; default to be 15 minutes)

xlab, ylab, xlim, ylim, main
they are passed to plot with reasonable default values

...
other arguments passed to plot

Value

invisible NULL

Author(s)

Yihui Xie

Examples

## see more examples in ?vanke1127
saveHTML({
    price.ani(vanke1127$price, vanke1127$time, lwd = 2)
}, img.name = "vanke1127", htmlfile = "vanke1127.html", title = "Stock prices of Vanke", 
    description = c("Barplots", "of the stock prices of Vanke Co. Ltd", "on 2009/11/27"))

[Package animation version 2.4 Index]