sample.system {animation}	R Documentation

Demonstration for the systematic sampling

Description

The whole sample frame is denoted by a matrix (nrow * ncol) in the plane, and the sample points with equal intervals are drawn out according to a random starting point. The points being sampled are marked by red circles.

Usage

sample.system(nrow = 10, ncol = 10, size = 15, p.col = c("blue", "red"), p.cex = c(1, 
    3))

Arguments

nrow
the desired number of rows of the sample frame.

ncol
the desired number of columns of the sample frame.

size
the sample size.

p.col, p.cex
different colors / magnification rate to annotate the population and the sample

Value

None (invisible NULL).

Author(s)

Yihui Xie

See Also

sample, sample.simple, sample.cluster, sample.ratio, sample.strat

Examples

oopt = ani.options(nmax = ifelse(interactive(), 50, 2))
par(mar = rep(1, 4), lwd = 2)

sample.system()

## HTML animation pages
saveHTML({
    ani.options(interval = 1, nmax = ifelse(interactive(), 30, 2))
    par(mar = rep(1, 4), lwd = 2)
    sample.system()
}, img.name = "sample.system", htmlfile = "sample.html", ani.height = 350, 
    ani.width = 500, title = "Demonstration of the systematic sampling", 
    description = "Sampling with equal distances.")

ani.options(oopt)

[Package animation version 2.4 Index]