vi.grid.illusion(animation)	R Documentation

Visual Illusions: Scintillating grid illusion and Hermann grid illusion

Description

A grid illusion is any kind of grid that deceives a person's vision. The two most common types of grid illusions are Hermann grid illusions and Scintillating grid illusions. This function provides illustrations for both illusions.

Usage


vi.grid.illusion(nrow = 8, ncol = 8, lwd = 8, cex = 3, 
    col = "darkgray", type = c("s", "h"))

Arguments

nrow 	 number of rows for the grid

ncol 	 number of columns for the grid

lwd 	 line width for grid lines

cex 	 magnification for points in Scintillating grid illusions

col 	 color for grid lines

type 	 type of illusions: 's' for Scintillating grid illusions and 'h' for Hermann grid illusions

Details

This is actually a static image; pay attention to the intersections of the grid and there seems to be some moving points (non-existent in fact).

Value

None.

Note

In fact there isn't any animation!

Author(s)

Yihui Xie

References

http://en.wikipedia.org/wiki/Grid_illusion

http://animation.yihui.name/animation:misc#visual_illusions

See Also

points, abline

Examples


# default to be Scintillating grid illusions 
vi.grid.illusion()

# set wider lines to see Hermann grid illusions 
vi.grid.illusion(type = "h", lwd = 22, nrow = 5, ncol = 5, 
    col = "white") 

[Package animation version 1.0-1 Index]