vi.lilac.chaser(animation)	R Documentation

Visual Illusions: Lilac Chaser

Description

Stare at the center cross for a few (say 30) seconds to experience the phenomena of the illusion.

Usage


vi.lilac.chaser(np = 16, col = "magenta", bg = "gray", p.cex = 7, c.cex = 5)

Arguments

np 	 number of points

col 	 color of points

bg 	 background color of the plot

p.cex 	 magnification of points

c.cex 	 magnification of the center cross

Details

Just try it out.

Value

None.

Note

In fact, points in the original version of `Lilac Chaser' are blurred, which is not implemented in this function. If you have any idea, please contact me.

Author(s)

Yihui Xie

References

http://en.wikipedia.org/wiki/Lilac_chaser

http://animation.yihui.name/animation:misc#lilac_chaser

See Also

points

Examples


oopt = ani.options(interval = 0.05, nmax = 20)
op = par(pty = "s") 
vi.lilac.chaser()

## Not run:  

# HTML animation page; nmax = 1 is enough! 
ani.options(ani.height = 480, ani.width = 480, outdir = getwd(), nmax = 1,
    interval = 0.05, title = "Visual Illusions: Lilac Chaser",
    description = "Stare at the center cross for a few (say 30) seconds
    to experience the phenomena of the illusion.")
ani.start()
par(pty = "s", mar = rep(1, 4))
vi.lilac.chaser()
ani.stop()

## End(Not run)
par(op) 
ani.options(oopt)

[Package animation version 1.0-1 Index]